Optimize
Create a Student class, that will store the details of the Student, below mentioned are the attributes of a student.
Reg. No.
Name
Email
Phone
Class
Department
The Student class will have the following methods:
A constructor to initialize the values of the Student
A method to print the Details of the Student.
A method (named 'toString()) that convert the student details to string, and can be used to write the student details to file.
In the main-method class, create an array of Student Class to hold maximum details of 100 Students.

In the menu-driven program, the menu options will have
Add a student
Adds the details of 1 student to the array of Student.
Search for a student
Search for the details of a student from the array of Student.
Searching can be done with Name or Register Number.
Update the details of a student
Update the student details based on the search based on name or register number.
Display all students
Displays the details of all students.
Save the details of each student in a file, with the student name as filename.